+++ GSClient/GSClient/GameObject.h
@@ -129,7 +129,7 @@ public:
 	void ReleaseUploadBuffers(); 
 public:
 	virtual void Animate(float fTimeElapsed);
-	virtual void Update(double fTimeElapsed);
+	virtual void Update(float fTimeElapsed);
 
 	virtual void OnPrepareRender();
 	virtual void Draw(ID3D12GraphicsCommandList* pd3dCommandList, CCamera* pCamera);
